+++ R/boe.R
@@ -14,6 +14,16 @@
 #' @export
 #' @examplesIf curl::has_internet()
 #' \donttest{
+#' # Bank Rate
+#' boe_data("IUDBEDR", "2015-01-01")
+#'
+#' # SONIA daily rate
+#' boe_data("IUDSOIA", "2015-01-01")
+#'
+#' # 10-year nominal par yield
+#' boe_data("IUDMNPY", "2015-01-01")
+#'
+#' # multiple series
 #' boe_data(c("IUMABEDR", "IUALBEDR"), "2015-01-01")
 #' }
 boe_data <- function(key, start_date, end_date = Sys.Date()) {
@@ -30,36 +40,37 @@ boe_data <- function(key, start_date, end_date = Sys.Date()) {
 }
 
 parse_boe_data <- function(xml) {
-  xml2::xml_ns_strip(xml)
-  series <- xml2::xml_children(xml)
-  res <- lapply(series, function(x) {
-    id <- xml2::xml_attr(x, "SCODE")
-    desc <- xml2::xml_attr(x, "DESC")
-    freq_name <- x |>
-      xml2::xml_find_first("./Cube[@FREQ_NAME]") |>
-      xml2::xml_attr("FREQ_NAME") |>
-      tolower()
-    attrs <- xml2::xml_find_all(x, "./Cube[@CAT_NAME]")
-    nms <- xml2::xml_attr(attrs, "CAT_NAME")
-    nms <- gsub(" ", "_", tolower(nms), fixed = TRUE)
-    attrs <- attrs |>
-      xml2::xml_attr("VAL_DESC") |>
-      setNames(nms) |>
-      as.list()
+  xml |>
+    xml2::xml_ns_strip() |>
+    xml2::xml_children() |>
+    map(function(x) {
+      id <- xml2::xml_attr(x, "SCODE")
+      desc <- xml2::xml_attr(x, "DESC")
+      freq_name <- x |>
+        xml2::xml_find_first("./Cube[@FREQ_NAME]") |>
+        xml2::xml_attr("FREQ_NAME") |>
+        tolower()
+      attrs <- xml2::xml_find_all(x, "./Cube[@CAT_NAME]")
+      nms <- xml2::xml_attr(attrs, "CAT_NAME")
+      nms <- gsub(" ", "_", tolower(nms), fixed = TRUE)
+      attrs <- attrs |>
+        xml2::xml_attr("VAL_DESC") |>
+        setNames(nms) |>
+        as.list()
 
-    vals <- xml2::xml_find_all(x, "./Cube[@TIME and @OBS_VALUE]")
-    date <- vals |> xml2::xml_attr("TIME") |> as.Date()
-    value <- vals |> xml2::xml_attr("OBS_VALUE") |> as.numeric()
-    dt <- data.table(date = date, id = id, value = value, description = desc, freq = freq_name)
-    setnames(dt, "id", "key")
-    dt[, (names(attrs)) := attrs]
-  })
-  rbindlist(res, fill = TRUE)
+      vals <- xml2::xml_find_all(x, "./Cube[@TIME and @OBS_VALUE]")
+      date <- as.Date(xml2::xml_attr(vals, "TIME"))
+      value <- as.numeric(xml2::xml_attr(vals, "OBS_VALUE"))
+      dt <- data.table(date = date, id = id, value = value, description = desc, freq = freq_name)
+      setnames(dt, "id", "key")
+      dt[, (names(attrs)) := attrs]
+    }) |>
+    rbindlist(fill = TRUE)
 }
 
 boe <- function(...) {
   request("https://www.bankofengland.co.uk/boeapps/database/_iadb-fromshowcolumns.asp") |>
-    req_user_agent("bbk (https://m-muecke.github.io/bbk)") |>
+    req_user_agent(bbk_user_agent()) |>
     req_url_query(..., xml.x = "yes", .multi = "comma") |>
     req_error(
       is_error = function(resp) {
@@ -67,6 +78,8 @@ boe <- function(...) {
       },
       body = boe_error_body
     ) |>
+    req_bbk_retry() |>
+    req_bbk_cache() |>
     req_perform() |>
     resp_body_xml()
 }