+++ R/utils.R
@@ -1,5 +1,5 @@
-`%||%` <- function(x, y) {
-  if (is.null(x)) y else x
+grepv <- function(...) {
+  grep(..., value = TRUE)
 }
 
 parse_date <- function(date, freq) {
@@ -13,7 +13,7 @@ parse_date <- function(date, freq) {
 }
 
 extract_metadata <- function(string, pattern, fixed = FALSE) {
-  x <- grep(pattern, string, value = TRUE, fixed = fixed)
+  x <- grepv(pattern, string, fixed = fixed)
   if (length(x) > 0L) {
     strsplit(x, ",", fixed = TRUE)[[1L]][[2L]]
   } else {